First pass at migrating to apollo-server 5
