+++ services/api/package.json
@@ -19,6 +19,8 @@
   "author": "amazee.io <hello@amazee.io> (http://www.amazee.io)",
   "license": "MIT",
   "dependencies": {
+    "@apollo/server": "^5.5.1",
+    "@as-integrations/express4": "^1.1.2",
     "@aws-sdk/client-s3": "^3.828.0",
     "@aws-sdk/lib-storage": "^3.837.0",
     "@aws-sdk/s3-presigned-post": "^3.929.0",
@@ -32,7 +34,6 @@
     "@types/node-cache": "^4.2.5",
     "amqp-connection-manager": "^4.1.14",
     "amqplib": "^0.10.7",
-    "apollo-server-express": "^3.13.0",
     "aws-sdk": "^2.378.0",
     "body-parser": "^1.18.2",
     "camelcase-keys": "^4.2.0",
@@ -43,11 +44,11 @@
     "elasticsearch": "^15.0.0",
     "express": "^4.16.1",
     "git-url-parse": "^13.1.0",
-    "graphql": "^16.10.0",
+    "graphql": "^16.11.0",
     "graphql-iso-date": "^3.5.0",
     "graphql-redis-subscriptions": "^2.7.0",
     "graphql-sse": "^2.5.4",
-    "graphql-subscriptions": "^1.0.0",
+    "graphql-subscriptions": "^2.0.0",
     "graphql-tools": "^9.0.18",
     "graphql-type-json": "^0.3.0",
     "graphql-upload": "^17.0.0",
